rename hsearch -> search.hybrid
